pyproject.toml
[project]
name = "batch-rag"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "streamlit (>=1.45.1,<2.0.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "chromadb (>=1.0.12,<2.0.0)",
    "torch (>=2.7.0,<3.0.0)",
    "transformers (>=4.52.4,<5.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "langchain-community (>=0.3.24,<0.4.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "langchain-openai (>=0.3.19,<0.4.0)",
    "langchainhub (>=0.1.21,<0.2.0)",
    "chromadb (>=1.0.12,<2.0.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "playwright (>=1.52.0,<2.0.0)",
    "openai (>=1.84.0,<2.0.0)",
    "faiss-cpu (>=1.11.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "torchvision (>=0.22.0,<0.23.0)"
]
packages = [
    { include = "batch_rag" }
]

[mypy]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
